# [Sparse] [llvm] [ir] Reuse up-to-date element lists Draft