--- a/taichi/backends/metal/aot_module_builder_impl.cpp
+++ b/taichi/backends/metal/aot_module_builder_impl.cpp
@@ -4,7 +4,7 @@
 
 #include "taichi/backends/metal/codegen_metal.h"
 
-#if !defined(__clang__) && defined(__GNUC__) && (__GNUC__ < 8)
+#if defined(__GNUC__) && (__GNUC__ < 8)
 // https://stackoverflow.com/a/45867491
 // !defined(__clang__) to make sure this is not clang
 // https://stackoverflow.com/questions/38499462/how-to-tell-clang-to-stop-pretending-to-be-other-compilers