--- a/taichi/runtime/llvm/node_dynamic.h
+++ b/taichi/runtime/llvm/node_dynamic.h
@@ -18,16 +18,18 @@ void Dynamic_activate(Ptr meta_, Ptr node_, int i) {
   auto node = (DynamicNode *)(node_);
   // We need to not only update node->n, but also make sure the chunk containing
   // element i is allocated.
-  meta->context->runtime->element_lists[meta->snode_id]->up_to_date = false;
   atomic_max_i32(&node->n, i + 1);
   int chunk_start = 0;
   auto p_chunk_ptr = &node->ptr;
   auto chunk_size = meta->chunk_size;
   while (true) {
     if (*p_chunk_ptr == nullptr) {
+      auto rt = meta->context->runtime;
+      for (int j = 0; j < meta->num_ch_snode; j++) {
+        rt->element_lists[meta->ch_snode_id[j]]->up_to_date = false;
+      }
       locked_task(Ptr(&node->lock), [&] {
         if (*p_chunk_ptr == nullptr) {
-          auto rt = meta->context->runtime;
           auto alloc = rt->node_allocators[meta->snode_id];
           *p_chunk_ptr = alloc->allocate();
         }
@@ -44,13 +46,14 @@ void Dynamic_activate(Ptr meta_, Ptr node_, int i) {
 void Dynamic_deactivate(Ptr meta_, Ptr node_) {
   auto meta = (DynamicMeta *)(meta_);
   auto node = (DynamicNode *)(node_);
-  auto list = meta->context->runtime->element_lists[meta->snode_id];
-  list->up_to_date = false;
   if (node->n > 0) {
+    auto rt = meta->context->runtime;
+    for (int j = 0; j < meta->num_ch_snode; j++) {
+      rt->element_lists[meta->ch_snode_id[j]]->up_to_date = false;
+    }
     locked_task(Ptr(&node->lock), [&] {
       node->n = 0;
       auto p_chunk_ptr = &node->ptr;
-      auto rt = meta->context->runtime;
       auto alloc = rt->node_allocators[meta->snode_id];
       while (*p_chunk_ptr) {
         alloc->recycle(*p_chunk_ptr);
@@ -70,9 +73,12 @@ i32 Dynamic_append(Ptr meta_, Ptr node_, i32 data) {
   auto p_chunk_ptr = &node->ptr;
   while (true) {
     if (*p_chunk_ptr == nullptr) {
+      auto rt = meta->context->runtime;
+      for (int j = 0; j < meta->num_ch_snode; j++) {
+        rt->element_lists[meta->ch_snode_id[j]]->up_to_date = false;
+      }
       locked_task(Ptr(&node->lock), [&] {
         if (*p_chunk_ptr == nullptr) {
-          auto rt = meta->context->runtime;
           auto alloc = rt->node_allocators[meta->snode_id];
           *p_chunk_ptr = alloc->allocate();
         }