Related issue = #451 

Lets call this a draft for now. Async mode not passing tests

I've noticed in my applications that the cost of listgen is far from negligible. The problem is we regenerate the list at the start of every kernel, regardless of whether any elements have actually been activated or deactivated. I see that Async mode takes steps to reduce this, but it doesn't go far enough.

This PR implements runtime checking of whether an element list is up-date-or not. If the list is up-to-date, meaning there have been no activation/deactivation of its elements, then listgen computation is skipped. This can lead to huge time savings especially on very sparse grids. Since the cost of the compute kernel decreases faster with more sparsity than the cost of listgen.  

Below shows timing of mgpcg_advanced.py with 256^3 grid. As you can see. List reuses led to about 10-15% cost saving. And this really isnt a great example since it can hardly be considered a sparse grid. 

Let me know what you think!

With list reuse
```
[Taichi] mode=development
[Taichi] preparing sandbox at /tmp/taichi-2a8qi4yy
[Taichi] <dev mode>, llvm 10.0.0, commit 65e5f154, linux, python 3.6.9
[Taichi] Starting on arch=x64
[Taichi] materializing...
iter 0, |residual|_2=6258.0782992864515
iter 1, |residual|_2=597.9025579055503
iter 2, |residual|_2=86.4125760568521
iter 3, |residual|_2=8.257598152374197
iter 4, |residual|_2=1.3431398647758535
iter 5, |residual|_2=0.1359300890908447
iter 6, |residual|_2=0.0272835500960537
iter 7, |residual|_2=0.00222295110914757
x64 Profiler
=========================================================================
[      %     total   count |      min       avg       max   ] Kernel name
[ 18.91%   1.497 s     32x |   42.765    46.773    50.170 ms] smooth_c22_0_kernel_26_struct_for
[ 18.69%   1.479 s     32x |   43.269    46.232    47.721 ms] smooth_c22_1_kernel_33_struct_for
[ 15.42%   1.221 s      8x |  147.583   152.595   159.815 ms] restrict_c18_0_kernel_54_struct_for
[  7.41%   0.587 s      8x |   68.274    73.348    75.595 ms] compute_Ap_c8_0_kernel_181_struct_for
[  5.14%   0.407 s     64x |    5.784     6.360     7.154 ms] smooth_c22_2_kernel_61_struct_for
[  5.10%   0.404 s     64x |    5.694     6.306     7.207 ms] smooth_c22_3_kernel_68_struct_for
[  3.68%   0.291 s      8x |   35.723    36.427    36.931 ms] update_r_c14_0_kernel_204_struct_for
[  3.09%   0.245 s      8x |   30.198    30.600    31.013 ms] prolongate_c20_2_kernel_159_struct_for
[  2.89%   0.229 s      1x |  228.907   228.907   228.907 ms] init_c24_0_kernel_3_range_for
[  2.59%   0.205 s      9x |   19.690    22.769    23.981 ms] reduce_c10_0_kernel_11_struct_for
[  2.42%   0.192 s      8x |   23.339    23.961    25.451 ms] update_p_c16_0_kernel_166_struct_for
[  2.21%   0.175 s      8x |   20.030    21.900    23.135 ms] reduce_c10_2_kernel_189_struct_for
[  2.19%   0.173 s      8x |   19.570    21.675    22.617 ms] reduce_c10_1_kernel_174_struct_for
[  1.92%   0.152 s      8x |   18.379    18.982    20.534 ms] restrict_c18_1_kernel_89_struct_for
[  1.83%   0.145 s      8x |   17.478    18.144    18.594 ms] update_x_c12_0_kernel_197_struct_for
[  1.30%   0.103 s    128x |    0.709     0.801     2.385 ms] smooth_c22_5_kernel_103_struct_for
[  1.29%   0.102 s    128x |    0.716     0.799     1.124 ms] smooth_c22_4_kernel_96_struct_for
[  1.26%   0.100 s      8x |   11.913    12.492    12.874 ms] fill_tensor_c28_0_kernel_19_struct_for
[  0.56%   0.044 s    400x |    0.098     0.110     0.421 ms] smooth_c22_7_kernel_138_struct_for
[  0.55%   0.044 s    400x |    0.097     0.109     0.225 ms] smooth_c22_6_kernel_131_struct_for
[  0.36%   0.029 s      8x |    3.286     3.608     3.784 ms] prolongate_c20_1_kernel_152_struct_for
[  0.24%   0.019 s      8x |    2.218     2.383     2.867 ms] restrict_c18_2_kernel_124_struct_for
[  0.21%   0.016 s      9x |    0.000     1.818    16.360 ms] reduce_c10_0_kernel_9_listgen_S7dense
[  0.21%   0.016 s      8x |    0.000     2.029    16.234 ms] update_p_c16_0_kernel_164_listgen_S2dense
[  0.15%   0.012 s      8x |    0.061     1.443     1.699 ms] fill_tensor_c28_1_kernel_40_struct_for
[  0.14%   0.011 s      8x |    0.022     1.429     2.033 ms] fill_tensor_c28_2_kernel_47_struct_for
[  0.05%   0.004 s      8x |    0.431     0.452     0.498 ms] prolongate_c20_0_kernel_145_struct_for
[  0.03%   0.002 s     64x |    0.000     0.032     2.017 ms] smooth_c22_2_kernel_59_listgen_S11dense
[  0.02%   0.002 s      8x |    0.021     0.233     0.322 ms] fill_tensor_c28_3_kernel_75_struct_for
[  0.02%   0.002 s      8x |    0.019     0.205     0.258 ms] fill_tensor_c28_4_kernel_82_struct_for
[  0.01%   0.001 s    128x |    0.000     0.009     1.136 ms] smooth_c22_4_kernel_94_listgen_S15dense
[  0.01%   0.001 s      8x |    0.000     0.127     1.015 ms] fill_tensor_c28_1_kernel_38_listgen_S11dense
[  0.01%   0.001 s      8x |    0.000     0.109     0.873 ms] fill_tensor_c28_3_kernel_70_listgen_S14pointer
[  0.01%   0.001 s      1x |    0.715     0.715     0.715 ms] paint_c26_0_kernel_210_range_for
[  0.01%   0.001 s      8x |    0.000     0.081     0.645 ms] fill_tensor_c28_5_kernel_105_listgen_S18pointer
[  0.01%   0.001 s      9x |    0.000     0.070     0.625 ms] reduce_c10_0_kernel_6_listgen_S6pointer
[  0.01%   0.001 s      8x |    0.000     0.073     0.579 ms] update_p_c16_0_kernel_161_listgen_S1pointer
[  0.00%   0.000 s    400x |    0.000     0.001     0.368 ms] smooth_c22_6_kernel_129_listgen_S19dense
[  0.00%   0.000 s      8x |    0.021     0.043     0.050 ms] fill_tensor_c28_5_kernel_110_struct_for
[  0.00%   0.000 s      8x |    0.020     0.042     0.049 ms] fill_tensor_c28_6_kernel_117_struct_for
[  0.00%   0.000 s      1x |    0.218     0.218     0.218 ms] tensor_to_ext_arr_c30_0_kernel_211_range_for
[  0.00%   0.000 s      8x |    0.000     0.013     0.106 ms] fill_tensor_c28_3_kernel_73_listgen_S15dense
[  0.00%   0.000 s      8x |    0.000     0.010     0.081 ms] fill_tensor_c28_1_kernel_35_listgen_S10pointer
[  0.00%   0.000 s    400x |    0.000     0.000     0.001 ms] smooth_c22_7_kernel_132_serial
[  0.00%   0.000 s    400x |    0.000     0.000     0.001 ms] smooth_c22_6_kernel_125_serial
[  0.00%   0.000 s    400x |    0.000     0.000     0.001 ms] smooth_c22_6_kernel_126_listgen_S18pointer
[  0.00%   0.000 s     64x |    0.000     0.000     0.001 ms] smooth_c22_2_kernel_55_serial
[  0.00%   0.000 s     64x |    0.000     0.000     0.001 ms] smooth_c22_3_kernel_62_serial
[  0.00%   0.000 s    400x |    0.000     0.000     0.001 ms] smooth_c22_6_kernel_128_serial
[  0.00%   0.000 s    400x |    0.000     0.000     0.001 ms] smooth_c22_7_kernel_133_listgen_S18pointer
[  0.00%   0.000 s    128x |    0.000     0.000     0.003 ms] smooth_c22_5_kernel_97_serial
[  0.00%   0.000 s    400x |    0.000     0.000     0.001 ms] smooth_c22_7_kernel_137_serial
[  0.00%   0.000 s      8x |    0.000     0.002     0.015 ms] fill_tensor_c28_5_kernel_108_listgen_S19dense
[  0.00%   0.000 s    400x |    0.000     0.000     0.001 ms] smooth_c22_7_kernel_136_listgen_S19dense
[  0.00%   0.000 s    128x |    0.000     0.000     0.001 ms] smooth_c22_4_kernel_90_serial
[  0.00%   0.000 s     32x |    0.000     0.000     0.001 ms] smooth_c22_1_kernel_27_serial
[  0.00%   0.000 s    400x |    0.000     0.000     0.001 ms] smooth_c22_6_kernel_130_serial
[  0.00%   0.000 s     64x |    0.000     0.000     0.001 ms] smooth_c22_3_kernel_63_listgen_S10pointer
[  0.00%   0.000 s    400x |    0.000     0.000     0.001 ms] smooth_c22_7_kernel_134_serial
[  0.00%   0.000 s    400x |    0.000     0.000     0.001 ms] smooth_c22_7_kernel_135_serial
[  0.00%   0.000 s     64x |    0.000     0.000     0.001 ms] smooth_c22_2_kernel_56_listgen_S10pointer
[  0.00%   0.000 s    128x |    0.000     0.000     0.001 ms] smooth_c22_4_kernel_93_serial
[  0.00%   0.000 s    400x |    0.000     0.000     0.001 ms] smooth_c22_6_kernel_127_serial
[  0.00%   0.000 s    128x |    0.000     0.000     0.001 ms] smooth_c22_5_kernel_98_listgen_S14pointer
[  0.00%   0.000 s     64x |    0.000     0.000     0.001 ms] smooth_c22_2_kernel_58_serial
[  0.00%   0.000 s     64x |    0.000     0.000     0.001 ms] smooth_c22_3_kernel_66_listgen_S11dense
[  0.00%   0.000 s    128x |    0.000     0.000     0.001 ms] smooth_c22_4_kernel_91_listgen_S14pointer
[  0.00%   0.000 s    128x |    0.000     0.000     0.001 ms] smooth_c22_5_kernel_100_serial
[  0.00%   0.000 s     32x |    0.000     0.000     0.001 ms] smooth_c22_0_kernel_20_serial
[  0.00%   0.000 s     32x |    0.000     0.000     0.001 ms] smooth_c22_0_kernel_21_listgen_S6pointer
[  0.00%   0.000 s    128x |    0.000     0.000     0.001 ms] smooth_c22_5_kernel_99_serial
[  0.00%   0.000 s      9x |    0.000     0.001     0.004 ms] reduce_c10_0_kernel_4_serial
[  0.00%   0.000 s     25x |    0.000     0.000     0.001 ms] snode_reader_27_kernel_12_serial
[  0.00%   0.000 s    128x |    0.000     0.000     0.001 ms] smooth_c22_4_kernel_92_serial
[  0.00%   0.000 s     32x |    0.000     0.000     0.001 ms] jit_evaluator_0_kernel_0_serial
[  0.00%   0.000 s     32x |    0.000     0.000     0.001 ms] smooth_c22_1_kernel_30_serial
[  0.00%   0.000 s     32x |    0.000     0.000     0.001 ms] smooth_c22_1_kernel_28_listgen_S6pointer
[  0.00%   0.000 s      8x |    0.000     0.001     0.001 ms] prolongate_c20_2_kernel_156_serial
[  0.00%   0.000 s     32x |    0.000     0.000     0.001 ms] smooth_c22_0_kernel_24_listgen_S7dense
[  0.00%   0.000 s    128x |    0.000     0.000     0.001 ms] smooth_c22_5_kernel_101_listgen_S15dense
[  0.00%   0.000 s      8x |    0.000     0.001     0.001 ms] update_p_c16_0_kernel_160_serial
[  0.00%   0.000 s     64x |    0.000     0.000     0.001 ms] smooth_c22_3_kernel_65_serial
[  0.00%   0.000 s      8x |    0.000     0.000     0.001 ms] reduce_c10_1_kernel_167_serial
[  0.00%   0.000 s     32x |    0.000     0.000     0.001 ms] smooth_c22_1_kernel_31_listgen_S7dense
[  0.00%   0.000 s      8x |    0.000     0.000     0.001 ms] fill_tensor_c28_1_kernel_37_serial
[  0.00%   0.000 s     32x |    0.000     0.000     0.001 ms] smooth_c22_0_kernel_25_serial
[  0.00%   0.000 s      8x |    0.000     0.000     0.001 ms] update_x_c12_0_kernel_195_listgen_S2dense
[  0.00%   0.000 s      9x |    0.000     0.000     0.001 ms] reduce_c10_0_kernel_5_serial
[  0.00%   0.000 s    128x |    0.000     0.000     0.001 ms] smooth_c22_5_kernel_102_serial
[  0.00%   0.000 s     64x |    0.000     0.000     0.001 ms] smooth_c22_3_kernel_67_serial
[  0.00%   0.000 s      8x |    0.000     0.000     0.001 ms] fill_tensor_c28_1_kernel_34_serial
[  0.00%   0.000 s      8x |    0.000     0.000     0.001 ms] restrict_c18_0_kernel_48_serial
[  0.00%   0.000 s      8x |    0.000     0.000     0.001 ms] prolongate_c20_2_kernel_153_serial
[  0.00%   0.000 s      8x |    0.000     0.000     0.001 ms] prolongate_c20_0_kernel_143_listgen_S15dense
[  0.00%   0.000 s      8x |    0.000     0.000     0.001 ms] fill_tensor_c28_2_kernel_41_serial
[  0.00%   0.000 s    128x |    0.000     0.000     0.001 ms] smooth_c22_4_kernel_95_serial
[  0.00%   0.000 s     64x |    0.000     0.000     0.001 ms] smooth_c22_2_kernel_60_serial
[  0.00%   0.000 s     64x |    0.000     0.000     0.001 ms] smooth_c22_2_kernel_57_serial
[  0.00%   0.000 s      8x |    0.000     0.000     0.001 ms] update_p_c16_0_kernel_163_serial
[  0.00%   0.000 s      8x |    0.000     0.000     0.001 ms] reduce_c10_1_kernel_168_serial
[  0.00%   0.000 s      8x |    0.000     0.000     0.001 ms] compute_Ap_c8_0_kernel_179_listgen_S2dense
[  0.00%   0.000 s      8x |    0.000     0.000     0.001 ms] update_r_c14_0_kernel_198_serial
[  0.00%   0.000 s      8x |    0.000     0.000     0.001 ms] fill_tensor_c28_3_kernel_72_serial
[  0.00%   0.000 s      8x |    0.000     0.000     0.001 ms] update_x_c12_0_kernel_192_listgen_S1pointer
[  0.00%   0.000 s      8x |    0.000     0.000     0.001 ms] prolongate_c20_1_kernel_146_serial
[  0.00%   0.000 s     32x |    0.000     0.000     0.001 ms] smooth_c22_1_kernel_32_serial
[  0.00%   0.000 s      7x |    0.000     0.000     0.001 ms] snode_writer_25_kernel_205_serial
[  0.00%   0.000 s      8x |    0.000     0.000     0.001 ms] restrict_c18_0_kernel_49_listgen_S6pointer
[  0.00%   0.000 s      8x |    0.000     0.000     0.001 ms] restrict_c18_0_kernel_51_serial
[  0.00%   0.000 s      8x |    0.000     0.000     0.001 ms] reduce_c10_2_kernel_183_serial
[  0.00%   0.000 s      8x |    0.000     0.000     0.001 ms] reduce_c10_2_kernel_184_listgen_S1pointer
[  0.00%   0.000 s      8x |    0.000     0.000     0.001 ms] prolongate_c20_2_kernel_157_listgen_S7dense
[  0.00%   0.000 s      8x |    0.000     0.000     0.001 ms] fill_tensor_c28_0_kernel_17_listgen_S7dense
[  0.00%   0.000 s      8x |    0.000     0.000     0.001 ms] reduce_c10_2_kernel_185_serial
[  0.00%   0.000 s      8x |    0.000     0.000     0.001 ms] restrict_c18_2_kernel_119_listgen_S14pointer
[  0.00%   0.000 s      9x |    0.000     0.000     0.001 ms] reduce_c10_0_kernel_10_serial
[  0.00%   0.000 s      8x |    0.000     0.000     0.001 ms] compute_Ap_c8_0_kernel_178_serial
[  0.00%   0.000 s      8x |    0.000     0.000     0.001 ms] fill_tensor_c28_4_kernel_79_serial
[  0.00%   0.000 s      8x |    0.000     0.000     0.001 ms] fill_tensor_c28_3_kernel_69_serial
[  0.00%   0.000 s      8x |    0.000     0.000     0.001 ms] prolongate_c20_1_kernel_151_serial
[  0.00%   0.000 s      8x |    0.000     0.000     0.001 ms] fill_tensor_c28_6_kernel_112_listgen_S18pointer
[  0.00%   0.000 s      8x |    0.000     0.000     0.001 ms] update_x_c12_0_kernel_191_serial
[  0.00%   0.000 s      8x |    0.000     0.000     0.001 ms] restrict_c18_1_kernel_83_serial
[  0.00%   0.000 s      8x |    0.000     0.000     0.001 ms] restrict_c18_2_kernel_122_listgen_S15dense
[  0.00%   0.000 s      8x |    0.000     0.000     0.001 ms] fill_tensor_c28_5_kernel_104_serial
[  0.00%   0.000 s      8x |    0.000     0.000     0.001 ms] fill_tensor_c28_0_kernel_14_listgen_S6pointer
[  0.00%   0.000 s      8x |    0.000     0.000     0.001 ms] update_r_c14_0_kernel_203_serial
[  0.00%   0.000 s      8x |    0.000     0.000     0.001 ms] fill_tensor_c28_2_kernel_42_listgen_S10pointer
[  0.00%   0.000 s      8x |    0.000     0.000     0.001 ms] prolongate_c20_1_kernel_149_serial
[  0.00%   0.000 s      8x |    0.000     0.000     0.001 ms] reduce_c10_1_kernel_173_serial
[  0.00%   0.000 s      8x |    0.000     0.000     0.001 ms] compute_Ap_c8_0_kernel_175_serial
[  0.00%   0.000 s      8x |    0.000     0.000     0.001 ms] compute_Ap_c8_0_kernel_176_listgen_S1pointer
[  0.00%   0.000 s      9x |    0.000     0.000     0.001 ms] reduce_c10_0_kernel_7_serial
[  0.00%   0.000 s      8x |    0.000     0.000     0.001 ms] prolongate_c20_2_kernel_154_listgen_S6pointer
[  0.00%   0.000 s      8x |    0.000     0.000     0.001 ms] reduce_c10_2_kernel_182_serial
[  0.00%   0.000 s      8x |    0.000     0.000     0.001 ms] reduce_c10_1_kernel_170_serial
[  0.00%   0.000 s      8x |    0.000     0.000     0.001 ms] update_p_c16_0_kernel_165_serial
[  0.00%   0.000 s      9x |    0.000     0.000     0.001 ms] reduce_c10_0_kernel_8_serial
[  0.00%   0.000 s     20x |    0.000     0.000     0.001 ms] jit_evaluator_2_kernel_2_serial
[  0.00%   0.000 s      8x |    0.000     0.000     0.001 ms] prolongate_c20_1_kernel_147_listgen_S10pointer
[  0.00%   0.000 s      8x |    0.000     0.000     0.001 ms] prolongate_c20_1_kernel_148_serial
[  0.00%   0.000 s      8x |    0.000     0.000     0.001 ms] prolongate_c20_0_kernel_139_serial
[  0.00%   0.000 s      8x |    0.000     0.000     0.001 ms] fill_tensor_c28_2_kernel_44_serial
[  0.00%   0.000 s      8x |    0.000     0.000     0.001 ms] fill_tensor_c28_2_kernel_45_listgen_S11dense
[  0.00%   0.000 s      8x |    0.000     0.000     0.001 ms] restrict_c18_0_kernel_52_listgen_S7dense
[  0.00%   0.000 s      8x |    0.000     0.000     0.001 ms] fill_tensor_c28_1_kernel_36_serial
[  0.00%   0.000 s      8x |    0.000     0.000     0.001 ms] update_r_c14_0_kernel_201_serial
[  0.00%   0.000 s      8x |    0.000     0.000     0.001 ms] fill_tensor_c28_4_kernel_76_serial
[  0.00%   0.000 s     32x |    0.000     0.000     0.001 ms] smooth_c22_1_kernel_29_serial
[  0.00%   0.000 s      8x |    0.000     0.000     0.001 ms] restrict_c18_1_kernel_84_listgen_S10pointer
[  0.00%   0.000 s      8x |    0.000     0.000     0.001 ms] update_x_c12_0_kernel_194_serial
[  0.00%   0.000 s      8x |    0.000     0.000     0.001 ms] restrict_c18_1_kernel_86_serial
[  0.00%   0.000 s      8x |    0.000     0.000     0.001 ms] restrict_c18_1_kernel_87_listgen_S11dense
[  0.00%   0.000 s     32x |    0.000     0.000     0.001 ms] smooth_c22_0_kernel_23_serial
[  0.00%   0.000 s      8x |    0.000     0.000     0.001 ms] prolongate_c20_0_kernel_140_listgen_S14pointer
[  0.00%   0.000 s      8x |    0.000     0.000     0.001 ms] fill_tensor_c28_5_kernel_109_serial
[  0.00%   0.000 s      8x |    0.000     0.000     0.001 ms] fill_tensor_c28_6_kernel_111_serial
[  0.00%   0.000 s      8x |    0.000     0.000     0.001 ms] fill_tensor_c28_6_kernel_114_serial
[  0.00%   0.000 s      8x |    0.000     0.000     0.001 ms] fill_tensor_c28_0_kernel_13_serial
[  0.00%   0.000 s      2x |    0.000     0.000     0.001 ms] jit_evaluator_5_kernel_208_serial
[  0.00%   0.000 s      6x |    0.000     0.000     0.001 ms] jit_evaluator_4_kernel_207_serial
[  0.00%   0.000 s      8x |    0.000     0.000     0.001 ms] reduce_c10_2_kernel_187_listgen_S2dense
[  0.00%   0.000 s      8x |    0.000     0.000     0.001 ms] reduce_c10_2_kernel_186_serial
[  0.00%   0.000 s      8x |    0.000     0.000     0.000 ms] update_r_c14_0_kernel_199_listgen_S1pointer
[  0.00%   0.000 s      8x |    0.000     0.000     0.000 ms] fill_tensor_c28_5_kernel_106_serial
[  0.00%   0.000 s      1x |    0.000     0.000     0.000 ms] jit_evaluator_3_kernel_206_serial
[  0.00%   0.000 s      2x |    0.000     0.000     0.000 ms] jit_evaluator_6_kernel_209_serial
[  0.00%   0.000 s      8x |    0.000     0.000     0.000 ms] compute_Ap_c8_0_kernel_177_serial
[  0.00%   0.000 s      8x |    0.000     0.000     0.000 ms] update_r_c14_0_kernel_202_listgen_S2dense
[  0.00%   0.000 s      8x |    0.000     0.000     0.000 ms] update_r_c14_0_kernel_200_serial
[  0.00%   0.000 s      8x |    0.000     0.000     0.000 ms] reduce_c10_2_kernel_188_serial
[  0.00%   0.000 s      8x |    0.000     0.000     0.000 ms] update_x_c12_0_kernel_196_serial
[  0.00%   0.000 s      8x |    0.000     0.000     0.000 ms] snode_writer_23_kernel_190_serial
[  0.00%   0.000 s      8x |    0.000     0.000     0.000 ms] update_x_c12_0_kernel_193_serial
[  0.00%   0.000 s      8x |    0.000     0.000     0.000 ms] compute_Ap_c8_0_kernel_180_serial
[  0.00%   0.000 s      8x |    0.000     0.000     0.000 ms] restrict_c18_0_kernel_50_serial
[  0.00%   0.000 s      8x |    0.000     0.000     0.000 ms] restrict_c18_1_kernel_85_serial
[  0.00%   0.000 s      8x |    0.000     0.000     0.000 ms] fill_tensor_c28_4_kernel_81_serial
[  0.00%   0.000 s      8x |    0.000     0.000     0.000 ms] fill_tensor_c28_4_kernel_80_listgen_S15dense
[  0.00%   0.000 s      8x |    0.000     0.000     0.000 ms] fill_tensor_c28_4_kernel_78_serial
[  0.00%   0.000 s      8x |    0.000     0.000     0.000 ms] fill_tensor_c28_4_kernel_77_listgen_S14pointer
[  0.00%   0.000 s      8x |    0.000     0.000     0.000 ms] fill_tensor_c28_3_kernel_74_serial
[  0.00%   0.000 s      8x |    0.000     0.000     0.000 ms] fill_tensor_c28_3_kernel_71_serial
[  0.00%   0.000 s     64x |    0.000     0.000     0.000 ms] smooth_c22_3_kernel_64_serial
…