@KLozes 2021-06-25
> Thanks for proposing this nice feature! One quick feedback: I believe up_to_date needs to be protected by the lock

@k-ye No problem! I actually don't think there needs to be a lock. `up_to_date` is never read from and written to in the same kernel.` listgen` and `clear_list` only read it. `set_list_up_to_date` only sets its to true. In compute kernels, many threads may try to set it to false at the same time, but this is okay since they are all trying to set it to the same value. So it wont matter that some writes get skipped.