@xumingkuan 2021-06-26
Nice feature! The Async mode reduces the number of listgen (and clear list) tasks at compile-time, and IIUC this PR makes the listgen (and clear list) tasks "empty" (makes them not generate the lists) at Python runtime. This PR will greatly help Sync mode save listgen time. I wonder if this PR has a time-saving effect on Async mode: are there cases that we cannot remove a listgen task at compile-time, but we can remove it at runtime?

> I believe test_sfg also needs to updated the to reflect that list generation is now 3 kernels (clear_list, listgen, setList_up_to_date) instead of 2. I believe the only code changes will be in state_flow_graph.cpp.

Right! I believe the only code changes will be in `StateFlowGraph::optimize_listgen()` in state_flow_graph.cpp. Note that in  https://github.com/taichi-dev/taichi/pull/2462/files#diff-978eb161b6e2752f494149a0df549810f73113763e0d1736778dff5f9c796b3eR232 , inserting an output list state means that the task is the owner (the latest writer) of the list state. So IIUC the condition at https://github.com/taichi-dev/taichi/blob/bed3558cd78a58712fa029b6bd9bb1fc34a67df8/taichi/program/state_flow_graph.cpp#L437-L438 still holds.

In `StateFlowGraph::optimize_listgen()`, we enumerate pairs of listgen tasks A and B (supposing B is after A), and eliminate B if B and A are guaranteed to generate the same lists. After eliminating B, this part of code eliminates the clear list task (statement) corresponding to the listgen task B:
https://github.com/taichi-dev/taichi/blob/bed3558cd78a58712fa029b6bd9bb1fc34a67df8/taichi/program/state_flow_graph.cpp#L485-L514
, among which the core part is
https://github.com/taichi-dev/taichi/blob/bed3558cd78a58712fa029b6bd9bb1fc34a67df8/taichi/program/state_flow_graph.cpp#L495-L502
. We can write some similar code to eliminate the set_list_up_to_date task (statement) after eliminating the listgen task B.