@KLozes 2021-06-26
>  I wonder if this PR has a time-saving effect on Async mode: are there cases that we cannot remove a listgen task at compile-time, but we can remove it at runtime?


Yes! imagine a pde solver on an adaptive grid. Where kernel1-3 evolve the solution of the pde.  If I understand Async mode correctly, its should recognize kernel1-3 use the same element list. So it removes the extra listgens before running kernel 2 and 3. At each timestep when we run kernel1 one again, it will run listgen. But notice that grid is only adapted every 100 timesteps. So we really don't need to update the list every timestep. I believe this is only something that can be achieved by running the program and tracking whether the lists are up-to-date or not. 
``` python
 for n in range(10000): # 10000 timesteps
     kernel1()
     kernel2()
     kernel3()
     if n%100 == 0:
         adapt_grid()    
```