@xumingkuan 2021-06-29
> > > I wonder if this PR has a time-saving effect on Async mode: are there cases that we cannot remove a listgen task at compile-time, but we can remove it at runtime?
> > 
> > 
> > Yes! Imagine a pde solver on an adaptive grid, where kernel1-3 evolve the solution of the pde. If I understand Async mode correctly, its should recognize kernel1-3 use the same element list. So it removes the extra listgens before running kernel 2 and 3. At each timestep when we run kernel1 one again, it will run listgen. But notice that grid is only adapted every 100 timesteps. So we really don't need to update the list every timestep. I believe this is only something that can be achieved by running the program and tracking whether the lists are up-to-date or not.
> > ```python
> >  for n in range(10000): # 10000 timesteps
> >      kernel1()
> >      kernel2()
> >      kernel3()
> >      if n%100 == 0:
> >          adapt_grid()    
> > ```
> 
> How about we add this as a test and use `ti.get_kernel_stats` for verification?

If this is to be added as a test, we may need to modify it a bit: note that if `if n%100 == 0` is in Python, the Async mode will be able to know if `adapt_grid()` will be executed or not, at Taichi compile-time (Python runtime).

So we may need to wrap the code in a Taichi kernel. However, inside a Taichi kernel, only the top-level for loop corresponds to an offloaded task. So maybe we cannot use `for n in range(10000)`...

An example I can think of now that the Async mode cannot perform the optimization is:
```python
@ti.kernel
def foo(n: ti.i32):
  func1()
  if n % 2 == 0:
    adapt_grid()
  func2()
```
where `n` is an argument with value `1` so that Taichi cannot know at compile-time if `adapt_grid()` will be executed.