@k-ye 2021-06-29
We had some discussion and wonder if you could also add a config flag (e.g. `enable_runtime_listgen_opt`) to conditionally do this optimization? The rationale is that this might introduce some runtime overhead, each time we activate/deactivate something. We may utilize this flag for some ablation study in the future :-)